Finnally remembering to turn this in.
I did this toolbox awhile ago and I also expanded on it for MP5. For MP5 I
I hid pictures within other pictures. I changed the last 3 bits of each rgb
band into the first three signfiicant bits of the hidden picture. The final
picture is a bit distorted but it still works great.

This toolbox came in handy during Google Games today. One of the clues
was a picture and I ran it through the origional toolbox program and got the
answer!!!

Thank you Soft Des!